requirements.txt
async-timeout==5.0.1       # Library for asynchronous timeouts (used in async programming)
blinker==1.9.0             # Fast, simple object-to-object and broadcast signaling (used for decoupling components)
cachelib==0.13.0           # A collection of cache backends for Python (used for caching)
certifi==2025.1.31         # Provides Mozilla's CA Bundle (used for verifying SSL certificates)
cffi==1.17.1               # Foreign Function Interface for Python calling C code (used for interacting with C libraries)
charset-normalizer==3.4.1  # Character set detection and normalization (used for handling text encodings)
click==8.1.8               # Command-line interface creation kit (used for building command-line tools)
colorama==0.4.6            # Cross-platform colored terminal text (used for adding color to console output)
Deprecated==1.2.18         # Library for deprecating code (used for marking code as deprecated)
Flask==3.1.0               # Micro web framework for Python (used for building web applications)
Flask-Limiter==3.11.0      # Rate limiting for Flask applications (used for controlling request rates)
Flask-Session==0.8.0       # Adds server-side session support to Flask (used for managing user sessions)
Flask-WTF==1.2.2           # Integrates WTForms with Flask (used for handling web forms)
gunicorn==23.0.0           # WSGI HTTP server (used for deploying Python web applications)
idna==3.10                 # Internationalized Domain Names in Applications (used for handling international domain names)
importlib_metadata==8.6.1  # Accessing package metadata (used for reading package information)
itsdangerous==2.2.0        # Various helpers to pass trusted data (used for signing data)
Jinja2==3.1.6              # Templating engine for Python (used for creating HTML templates)
limits==4.2                # Library for rate limiting (used by Flask-Limiter)
markdown-it-py==3.0.0      # Markdown parser (used for parsing Markdown text)
MarkupSafe==3.0.2          # Escapes text to prevent injection attacks (used by Jinja2)
mdurl==0.1.2               # URL parsing library (used by markdown-it-py)
msgspec==0.19.0            # Fast serialization and validation library.
ordered-set==4.1.0         # A mutable set that remembers the order of its entries.
packaging==24.2            # Core utilities for Python packages (used for handling package versions)
pycparser==2.22             # C parser for Python (used by cffi)
Pygments==2.19.1           # Syntax highlighter (used for highlighting code)
PyNaCl==1.5.0              # Python binding to the libsodium library (used for cryptography)
redis==5.2.1               # Python client for Redis (used for interacting with Redis)
requests==2.32.3           # HTTP library for Python (used for making HTTP requests)
rich==13.9.4               # Library for rich text and beautiful formatting in the terminal.
typing_extensions==4.13.0  # Backports and experimental type hints (used for type hinting)
urllib3==2.3.0             # HTTP client library (used by requests)
Werkzeug==3.1.3            # WSGI utility library (used by Flask)
wrapt==1.17.2              # Module wrapper (used for creating decorators)
WTForms==3.2.1             # Flexible forms validation and rendering library (used for handling web forms)
zipp==3.21.0               # Backport of pathlib-compatible zipfile object wrapper (used for working with zip files)
bcrypt==4.1.2              # Password hashing library.
